+++ test/cli/debug.t
@@ -7,7 +7,7 @@
   >     A.x
   > EOF
 
-  $ (ocamlformat --debug a.ml) 2>&1 | sed 's/\([^ ]*\.boxes\.ml$\)/<file>.boxes.ml/g'
+  $ (ocamlformat --debug a.ml) 2>&1 | sed -e 's/\([^ ]*\.boxes\.ml$\)/<file>.boxes.ml/g' -e 's/\([^ ]*\.ast$\)/<file>.ast/g'
   AST:
   Ptop_def
     [
@@ -63,6 +63,7 @@
         ]
     ]
   
+  [DEBUG] AST structure: <file>.ast
   AST:
   Ptop_def
     [
@@ -130,7 +131,7 @@
   > (* after let-binding *)
   > EOF
 
-  $ (ocamlformat --debug a.ml) 2>&1 | sed 's/\([^ ]*\.boxes\.ml$\)/<file>.boxes.ml/g'
+  $ (ocamlformat --debug a.ml) 2>&1 | sed -e 's/\([^ ]*\.boxes\.ml$\)/<file>.boxes.ml/g' -e 's/\([^ ]*\.ast$\)/<file>.ast/g'
   AST:
   Ptop_def
     [
@@ -236,6 +237,7 @@
         ]
     ]
   
+  [DEBUG] AST structure: <file>.ast
   AST:
   Ptop_def
     [